+++ dev/examples/draggable.tsx
@@ -6,11 +6,6 @@ const styleA = {
     height: 300,
     background: "blue",
 }
-const styleB = {
-    width: 100,
-    height: 100,
-    background: "red",
-}
 
 export const App = () => {
     return <motion.div drag style={styleA} />