+++ jest.setup.ts
@@ -1,6 +1,6 @@
 import "jest-dom/extend-expect"
 // Get fireEvent from the native testing library
-// because react-testing-library one switches out mouseEnter and mouseLeave
-import { fireEvent } from "dom-testing-library"
+// because @testing-library/react one switches out mouseEnter and mouseLeave
+import { fireEvent } from "@testing-library/dom"
 export const { mouseEnter, mouseLeave } = fireEvent
-import "react-testing-library/cleanup-after-each"
+import "@testing-library/react/cleanup-after-each"