+++ src/behaviours/utils/block-viewport-scroll.ts
@@ -1,16 +1,17 @@
-import { safeWindow } from "../../events/utils/window"
-
 let isViewportScrollBlocked = false
+const isBrowser = typeof window !== "undefined"
 
-safeWindow.addEventListener(
-    "touchmove",
-    (event: TouchEvent) => {
-        if (isViewportScrollBlocked) {
-            event.preventDefault()
-        }
-    },
-    { passive: false }
-)
+if (isBrowser) {
+    window.addEventListener(
+        "touchmove",
+        (event: TouchEvent) => {
+            if (isViewportScrollBlocked) {
+                event.preventDefault()
+            }
+        },
+        { passive: false }
+    )
+}
 
 export const blockViewportScroll = () => (isViewportScrollBlocked = true)
 export const unblockViewportScroll = () => (isViewportScrollBlocked = false)