+++ src/events/types.ts
@@ -47,7 +47,10 @@ export interface EventInfo {
     point: Point
 }
 
-export type EventHandler = (event: Event, info: EventInfo) => void
+export type EventHandler = (
+    event: MouseEvent | TouchEvent | PointerEvent,
+    info: EventInfo
+) => void
 
 export type ListenerControls = [() => void, () => void]