+++ src/gestures/utils/is-touch-event.ts
@@ -1,6 +1,0 @@
-export function isTouchEvent(
-    event: MouseEvent | TouchEvent
-): event is TouchEvent {
-    const hasTouches = !!(event as TouchEvent).touches
-    return hasTouches
-}