+++ src/motion/functionality/dom.tsx
@@ -11,7 +11,7 @@ import { Drag } from "./drag"
 import { parseDomVariant } from "../../dom/parse-dom-variant"
 import { MotionValuesMap } from "../../motion/utils/use-motion-values"
 import { resolveCurrent } from "../../value/utils/resolve-values"
-import { Position } from "./position"
+import { Layout } from "./layout"
 import { isValidMotionProp } from "../utils/valid-prop"
 import { getAnimationComponent } from "./animation"
 
@@ -29,12 +29,19 @@ function stripMotionProps(props: MotionProps) {
 
 const buildSVGProps = (values: MotionValuesMap, style: CSSProperties) => {
     const motionValueStyles = resolveCurrent(values)
-    const props = buildSVGAttrs(motionValueStyles)
+    const props = buildSVGAttrs(
+        motionValueStyles,
+        undefined,
+        undefined,
+        undefined,
+        undefined,
+        false
+    )
     props.style = { ...style, ...props.style } as any
     return props
 }
 
-const functionalityComponents = [Position, Drag, Gestures]
+const functionalityComponents = [Layout, Drag, Gestures]
 const numFunctionalityComponents = functionalityComponents.length
 
 /**
@@ -51,6 +58,7 @@ export function createDomMotionConfig<P = MotionProps>(
     return {
         renderComponent: (ref, style, values, props, isStatic) => {
             const forwardProps = isDOM ? stripMotionProps(props) : props
+
             const staticVisualStyles = isSVG
                 ? buildSVGProps(values, style)
                 : { style: buildStyleAttr(values, style, isStatic) }