+++ src/utils/use-unmount-effect.ts
@@ -1,0 +1,5 @@
+import { useEffect } from "react"
+
+export function useUnmountEffect(callback: () => void) {
+    return useEffect(() => () => callback(), [])
+}